+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.0.4] - 2022-XX-XX
+* Maintenance
+  * Reduce duplication of code in instrument modules
+
 ## [0.0.3] - 2022-05-18
 * Include flake8 linting of docstrings and style in Github Actions
 * Include Windows tests in Github Actions